@@ -12,7 +12,8 @@
     }
   ],
   "require": {
-    "laravel/framework": "^7.0"
+    "laravel/framework": "^8.0",
+    "msurguy/honeypot": "^1.0"
   },
   "autoload": {
     "files": [
